server4_restructured/server4_restructured/streamlit-docker/requirements.txt
streamlit>=1.35.0
requests
python-dotenv
pandas
pyyaml
